feat: functional prune command
Adds a `tk prune` command, that does what it promises. Prune is
functionally complete now.
