+++ pkg/spec/v1alpha1/config.go
@@ -42,5 +42,4 @@ type Spec struct {
 	APIServer    string `json:"apiServer"`
 	Namespace    string `json:"namespace"`
 	DiffStrategy string `json:"diffStrategy,omitempty"`
-	Prune        bool   `json:"prune,omitempty"`
 }
